db_service/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
asyncpg
python-dotenv
python-dateutil
pydantic
pydantic-settings